Moved the hint functions to util
